requirements.txt
streamlit
langchain-groq
python-dotenv
langchain-core